Share BeatmapSet cover sprite code.
